requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
pyTelegramBotAPI==4.15.0
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.32.3
replicate==0.23.1
gunicorn==21.2.0
pandas==2.2.2
gevent==23.9.1
sqlalchemy==2.0.23
werkzeug==2.3.6
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3
bcrypt
psutil
pyjwt
